Split modules up. Added main py to allow easier imports.
